New historic displaying + confirmation modal

+++ src/storageHistory.js
@@ -1,15 +1,15 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { Table } from 'react-bootstrap';
+import { Table, Tooltip, OverlayTrigger, Button, Modal } from 'react-bootstrap';
 
 export default class storageHistory extends React.Component {
 
     static cutQuery(query) {
         let res = query;
-        if (query.length <= 200) {
+        if (query.length <= 100) {
             return res;
         }
-        res = res.substring(0, 200);
+        res = res.substring(0, 100);
         if (res.substring(res.Length, 1) !== ' ') {
             const LastSpace = res.lastIndexOf(' ');
             if (LastSpace !== -1) {
@@ -23,50 +23,70 @@ export default class storageHistory extends React.Component {
     constructor(props) {
         super(props);
         this.localStorage = JSON.parse(window.localStorage.getItem('dlISTEX'));
+        this.state = {
+            showConfirm: false,
+        };
     }
+
     render() {
+        const editTooltip = (
+            <Tooltip data-html="true" id="dlTooltip">
+                Editer cette requête
+            </Tooltip>);
+        const dlTooltip = (
+            <Tooltip data-html="true" id="dlTooltip">
+                Télécharger cette requête
+            </Tooltip>);
         this.localStorage = JSON.parse(window.localStorage.getItem('dlISTEX'));
-        this.nomColonnes = this.props.nomColonnes.split(',');
-        this.tableauColonnes = [];
-        for (let j = 0; j < this.nomColonnes.length; j += 1) {
-            this.tableauColonnes[j] = (<th key={this.nomColonnes[j]}>{this.nomColonnes[j]}</th>);
+        this.columnNames = this.props.columnNames.split(',');
+        this.columnTab = [];
+        for (let j = 0; j < this.columnNames.length; j += 1) {
+            const onOneLign = this.columnNames[j].replace(' ', '\xa0');
+            this.columnTab[j] = (<th key={this.columnNames[j]}>{onOneLign}</th>);
         }
 
-        this.tableauHistorique = [];
-        this.vraiformatLigne = [];
+        this.hyistoryTab = [];
+        this.correctformatLine = [];
         if (window.localStorage && this.localStorage) {
-            const ancien = this.localStorage;
-            for (let i = 0; i < ancien.length; i += 1) {
-                const formatLigne = this.localStorage[i].formats;
-                this.vraiformatLigne[i] = [];
-                for (let x = 0; x < formatLigne.length; x += 1) {
-                    this.vraiformatLigne[i][x] = (<div key={formatLigne[x]}>{formatLigne[x]}</div>);
+            for (let i = 0; i < this.localStorage.length; i += 1) {
+                const formatLine = this.localStorage[i].formats;
+                this.correctformatLine[i] = [];
+                for (let x = 0; x < formatLine.length; x += 1) {
+                    this.correctformatLine[i][x] = (<div key={formatLine[x]}>{formatLine[x]}</div>);
                 }
-                this.tableauHistorique[i] = (
+                this.hyistoryTab[i] = (
                     <tr key={`table ${i}`}>
                         <td>{i + 1}</td>
                         <td>{new Date(this.localStorage[i].date).toUTCString()}</td>
-                        <td>{this.vraiformatLigne[i]}</td>
+                        <td>{this.correctformatLine[i]}</td>
                         <td>{this.localStorage[i].size}</td>
                         <td>{storageHistory.cutQuery(this.localStorage[i].q)}</td>
-                        <td><button
-                            type="button" className="btn-sm"
-                            onClick={() => {
-                                    // this.interpretURL(this.localStorage[i].url);
-                                window.location = this.localStorage[i].url;
-                            }}
-                        >
-                            Editer cette requête
-                        </button>
+                        <td className="transparent-td">
+                            <OverlayTrigger
+                                placement="top"
+                                overlay={editTooltip}
+                                onClick={() => {
+                                        // this.interpretURL(this.localStorage[i].url);
+                                    window.location = this.localStorage[i].url;
+                                }}
+                            >
+                                <span
+                                    role="button" className="glyphicon glyphicon-edit"
+                                />
+                            </OverlayTrigger>
                         </td>
-                        <td><button
-                            type="button" className="btn-sm"
-                            onClick={() => {
-                                window.location = `${this.localStorage[i].url}&download=true`;
-                            }}
-                        >
-                            Télécharger cette requête
-                        </button>
+                        <td className="transparent-td">
+                            <OverlayTrigger
+                                placement="top"
+                                overlay={dlTooltip}
+                                onClick={() => {
+                                    window.location = `${this.localStorage[i].url}&download=true`;
+                                }}
+                            >
+                                <span
+                                    role="button" className="glyphicon glyphicon-download-alt"
+                                />
+                            </OverlayTrigger>
                         </td>
                     </tr>);
             }
@@ -74,32 +94,64 @@ export default class storageHistory extends React.Component {
 
         return (
             <div className="history">
-                <Table striped bordered condensed hover>
+                <Table responsive condensed hover>
                     <thead>
                         <tr>
-                            {this.tableauColonnes}
+                            {this.columnTab}
                         </tr>
                     </thead>
                     <tbody>
-                        {this.tableauHistorique}
+                        {this.hyistoryTab}
                     </tbody>
                 </Table>
-                <button
-                    type="button" className="btn-sm"
+                <Button
+                    bsStyle="danger"
                     onClick={() => {
-                        window.localStorage.clear();
                         this.setState({
-                            update: true,
+                            showConfirm: true,
                         });
                     }}
                 >
                 Supprimer l&apos;historique
-            </button>
+            </Button>
+                <Modal bsSize="small" show={this.state.showConfirm} onHide={this.close}>
+                    <Modal.Header>
+                        <Modal.Title>Confirmation</Modal.Title>
+                    </Modal.Header>
+
+                    <Modal.Body>
+                        Etes-vous sûr de vouloir supprimer l&apos;historique de vos téléchargements
+                    </Modal.Body>
+
+                    <Modal.Footer>
+                        <Button
+                            onClick={() => {
+                                this.setState({
+                                    showConfirm: false,
+                                });
+                            }}
+                        >
+                            Annuler
+                        </Button>
+                        <Button
+                            bsStyle="primary"
+                            onClick={() => {
+                                window.localStorage.clear();
+                                this.setState({
+                                    showConfirm: false,
+                                    update: true,
+                                });
+                            }}
+                        >
+                            Confirmer
+                        </Button>
+                    </Modal.Footer>
+                </Modal>
             </div>
         );
     }
 }
 
 storageHistory.propTypes = {
-    nomColonnes: PropTypes.string.isRequired,
+    columnNames: PropTypes.string.isRequired,
 };
